se muestra los pokemones en nuestra pantalla principal, fatlta dar funcionalidad al boton ver mas
