+++ src/index.html
@@ -86,7 +86,9 @@
         </nav>
     </header>
     <main>
-        <p id="mypokemon"></p>
+        <article id="numpokemon">
+            <!-- html medianete js -->
+        </article>
         <button id="btn-show" class="boton">VER +</button>
         <!-- Ficha Tecnica -->
         <article id="DatasheetPokemon">
